fix bugs in argparse

+++ data_generation_process/parallel_data_generation.py
@@ -35,9 +35,9 @@ def get_dgp_args():
                         help='highest frequency to generate. [default: 523.25 hz]')                  
     parser.add_argument('--major_prop', type=float, default=0.5,
                         help='proportion of major sample in generated data. [default: 0.5]')
-    parser.add_argument('--nb_cpu', type=float, default=2,
+    parser.add_argument('--nb_cpu', type=int, default=2,
                         help='number of cpu to generate data. [default: 2]')
-    parser.add_argument('--seed', type=float, default=0,
+    parser.add_argument('--seed', type=int, default=0,
                         help='random seed to determine the random melody sample. [default: 0]')
     parser.add_argument('--noise_level', type=float, default=0.0,
                         help='proportion of sample exposed to label noise. [default: 0.0]')
